secp256k1: Redefine in terms of v2 module.
This redefines the v1 secp256k1 module in terms of the v2 module by
making use of type aliases.  This will allow the v2 module types to be
used in place of the v1 types in other modules without requiring them to
undergo a major version bump.

Thus it updates the dcrec/secp256k1 v1 module dependencies and serves as
a base for dcrec/secp256k1/v1.0.3.

The updated direct dependencies in this commit are as follows:

- github.com/decred/dcrd/chaincfg/chainhash@v1.0.2
- github.com/decred/dcrd/dcrec/secp256k1/v2@v2.0.0

The full list of updated direct dependencies since the previous
dcrec/secp256k1/v1.0.2 release are as follows:

- github.com/decred/dcrd/chaincfg/chainhash@v1.0.2
- github.com/decred/dcrd/dcrec/secp256k1/v2@v2.0.0