cpuminer: Improve speed stat tracking.
Currently, in order to allow both the normal continuous and discrete CPU
mining modes, a non-blocking channel send is used to update the number
of hashes per second since the speed monitor is not running or required
in discrete mode.

As a result, in the case the total hashes are already being updated by
another goroutine, the update will be ignored which can lead to
reporting a lower hash rate than is accurate.

In order to better handle the aforementioned case, this modifies the
code to introduce a separate struct with its own mutex to be used for
tracking the speed stats and uses a unique instance for the continuous
and discrete mining modes.  A separate instance is used to ensure the
discrete mining mode is unable to interfere with the continuous mode
stats.