gcs: Prevent empty data elements fp matches.
This prunes any empty/nil data elements from the input array to MatchAny
to prevent potential false positive matches of empty data.

The matching code is not used in consensus, so it is safe to modify
without a vote and for all filter versions.

The single match function already failed attempts to match an empty
element as intended, however, prior to this change, it was possible to
match an empty item in the multi-item matching path due to a false
positive.  This is not desirable since the matching behavior must be
consistent in both the single and multi-match cases.