server: Convert lifecycle to context.
This modifies the lifecycle of the server to use the expected pattern
for running subsystems based on contexts in the future.

In particular, this renames the Start function to Run and arranges for
it to block until the provided context is cancelled.  This is more
flexible for the caller since it can easily turn blocking code into
async code while the reverse is not true.

The new Run function ensures server waits for all goroutines that it
starts to shutdown before shutting down itself to help ensure an orderly
shutdown.

Finally, this removes the WaitForShutdown function from the server since
it is no longer needed.