rpcserver: Start separate internal package impl.
This is the first of a series of commits that aims to make progress
towards being able to split the rpcserver into a separate internal
package.

It consists of creating the package, moving the rpcserverPeer interface
to the new package as the Peer interface so it becomes rpcserver.Peer
and updating the code to use the moved interface.