multi: Introduce regentemplate command
This adds the regentemplate comand to the node, which asks it to
regenerate is block template.

The node now maintains a background template generator, with various
timeouts to enable efficient generation of block templates in a
production setting.

However, external users might want to force a new template to be
generated after new regular transactions have been received but before
the regular regen timeout (currently 30 seconds) has elapsed. This is
true specially for external simnet tests that publish transactions and
rely on them being mined shortly after the node receives them.

The most flexible solution found was to add a new regentemplate command
which rpc clients can issue to force a node to regenerate the latest
template with any new transactions.