mining: Introduce PriorityInputser interface.
This introduces a new interface named PriorityInputser and updates the
functions that currently take a pointer to a blockchain.UtxoViewpoint
struct to accept the interface instead.

This removes the tight coupling between the two packages at the API
boundary and allows callers to easily provide custom values if desired.

It also updates the blockchain.UtxoViewpoint struct to implement the
interface.