rpcserver: Move error check for generate RPC.
This moves the error check for an attempt to call the generate RPC when
on a network that there is virtually no chance of mining a block with
the CPU into the RPC server where it more naturally belongs.  The caller
of the CPU miner should be the one to determine if it wants to allow
mining or not.  While here, use a more accurate RPC error code of
ErrDifficulty instead of ErrInternal.

This change is a step towards being able to separate the CPU mining code
into its own subpackage.