blockchain: Remove unconfigurable chain var.
This removes the main chain block size instance variable since it is
merely set to a constant and is not configurable anyway.  While here, it
also moves the prune height calculation outside of the loop and renames
the variables related to the main chain block cache to be more
consistent with the rest of the code.