blockchain: Rename KnownValid to HasValidated.
This renames the KnownValid and associated status flag of statusValid to
HasValidated and statusValidated, respectively in order to better
clarify the behavior.

While this distinction doesn't currently make much difference the
intention in the future to decouple the block process and download logic
at which point the distinction will become more important.