multi: Introduce initial sync min known chain work.
This adds a new parameter to the chain parameters to allow a minimum
total known cumulative amount of chain work to be specified as of a
given point in time and updates the check which determines if the
blockchain believes it is current to make use of the value instead of
the height of the final checkpoint.  In other words, the chain is not
considered synced if the best chain does not yet have the amount of work
specified.

This approach is preferred because it easier to maintain, does not rely
on trusted known good checkpoints, is valid regardless of the order that
blocks are received and processed, and, in general, is a more objective
value that can't be invalidated by a chain reorganization.

It also adds initial values for the main and test networks as follows:

mainnet: 0x0000000000000000000000000000000000000000000ae01920a7ee4b769cc620
testnet: 0x0000000000000000000000000000000000000000000000005df2701ec6263182