blockchain: Separate processing order tests.
Currently, the full block tests, which are intended to test the
consensus rules include some tests that are related to
implementation-dependent processing specifics as opposed to consensus
rules.  For example, rejection of duplicate blocks, accepting orphan
blocks, and handling of reorgs to invalid blocks when processed out of
order.

In order to better separate the consensus rules from the processing
aspects, this introduces a new test function specifically for testing
the process order logic and moves the related tests out of the full
block consensus tests into the new function.