blockchain: Support hdr checkpoints and simplify.
This reworks the checkpoint logic in blockchain to greatly simplify it
and pave the way to support checkpoints based on only the headers being
available.

Previously there was more complex logic that attempted to find the most
recent checkpoint that is already available in the downloaded portion of
the block chain on demand along with the next checkpoint that is
expected.  Instead, the new logic this introduces updates the most
recent checkpoint as the blocks are processed as needed and finds the
most recent checkpoint once during initialization.

This approach will eventually make it possible to update the most recent
checkpoint based on headers alone since it is no longer tightly tied to
the active best chain which will be important when decoupling the
connection code from the download logic.