rpcwebsocket: Remove client from missed maps
This removes outstanding references of a wsClient quit chan to the
notification maps in notificationHandler() upon client disconnection.

Previously, the left over references caused the wsClient and its
assotiated wsClientFilter struct to not be GCd after a wallet
disconnected from the node, which leaked memory.

This is specially relevant for setups that involve wallets with very
large number of addresses.