chaincfg: Remove chainec package.
The chainec package was originally intended to generalize the process of
adding alternate forms of crypto to the project, however, the code has
since been updated to avoid using it because it is not really a good
idea for several reasons, some of which are discussed below.  As a
result, since the package is no longer used, this removes it.

Crypto is one of the most expensive components of the overall entire
functionality of the daemon, so using a generic API that requires a lot
of otherwise unnecessary expensive conversions in and out of field and
group elements solely for the purposes of making an interface happy only
slows down the code in the most performance critical portions.

Further, a generic interface at those junction points significantly
hinders a lot of potential optimizations that can otherwise be made such
as avoiding field inversions back into affine coordinates during the
ECDSA signing process.

Another issue with the package is that, as the name indicates, it is
geared specifically towards elliptic curve crypto, however, it is highly
likely crypto based on other primitives, such as lattices, will be
introduced in the future which would be impossible to realistically fit
within the confines of an interface built for EC.

Finally, a particularly nasty consequence that could be very easy to
miss when using a generic interface is that there are several advanced
techniques used in various areas that make assumptions about the
specific properties of the underlying curve in use that if accidentally
used with a different curve would break in potentially subtle ways.

For example, the algorithm used for hierarchical deterministic key
generation specifically relies on the fact the private keys of the
secp256k1 curve form an affine linear space.  On the other hand,
Curve25519, used in the Ed25519 signature scheme, does not have that
property.  Moreover, Ed25519 uses a key generation scheme that is not
compatible either.  It is possible to make the derivation process work
with Ed25519, however, it does require modifications.  In other words,
blindly using a generic chainec interface with the derivation is one
example where it could easily lead to improper results that might not
even immediately be detectable since it sometimes works.  There are
other similar cases and therefore the entire notion of using a generic
interface really must be avoided.