secp256k1: Make params standalone.
This separates the params from the curve used to satisfy the
crypto/elliptic stdlib interface by creating a new type which embeds the
cyprto/elliptic.Params and adds the additional params provided by the
package.  This will still allow callers access to this params without
internally needing to go through the curve interface used to provide a
stdlib adaptor.

This is continued work towards making the internals of the package
entirely independent of the crypto/elliptic and crypto/ecdsa stdlib
interfaces.