secp256k1: Make public key independent type.
This modifies the PublicKey type to be defined as its own struct as
opposed of in terms of an ecdsa.PublicKey and updates the code
accordingly.

For now, it essentially keeps the same format with the exception of
embedding the elliptic.Curve, but being its own type will allow for it
to eventually be made opaque and thus provide greater optimizations.

This is continued work towards making the internals of the package
entirely independent of the crypto/elliptic and crypto/ecdsa stdlib
interfaces.