secp256k1: Follow RFC6979 for too large nonce data.
This modifies the NonceRFC6979 function to follow the RFC for data that
is larger than the expected size.  In particular, section 2.3.2
specifically says that the sequence is first _truncated_ or expanded to
the expected length.  The exact wording is:

if qlen < blen, then the qlen leftmost bits are kept, and subsequent
bits are discarded

The current code is dropping the leftmost bits (most significant in big
endian) instead of the rightmost.

While here, this also switches to a byte slice for the key param instead
of a big integer to avoid requiring big integers which will be going
away in the future.

This also adds a bunch of additional tests to ensure the expected
behavior in all scenarios with malformed data as well as including tests
for the extra data and version portion which were previously lacking.

In practice, nothing in the code base is calling the function with too
many bytes, so this has no realistic effect on anything, but the
function should actually follow the spec.