secp256k1: Return new scalar from NonceRFC6979.
This modifies the NonceRFC6979 function to return the new ModNScalar
type instead of a big.Int to move closer towards big integer removal in
the primary code paths.

While here, this also adds a few more explicit zeros of memory when
signing which is good practice when dealing with key material.