secp256k1: Use new mod n scalar in ec mults.
This modifies the internal scalarMultJacobian and scalarBaseMultJacobian
functions to accept the new ModNScalar type instead of a raw byte slices
and updates the code accordingly.

It also moves the moduloReduce function from curve.go to
ellipticadaptor.go since it is only required in the adaptor code now.

This is work towards eventually using the new more efficient mod n
scalar throughout.