rpcserver: Refactor and update documentation
For createrawsstx explicitly state that all values are in atoms in help
messages and the documentation.

For createrawssrtx explicitly state that all values are in coins in help
messages and the documentation. Remove range loop for only one input.

For createrawtransaction explicitly state that all values are in coins in
help messages and the documentation. Move conversion to atoms above the
check for max amount and check after converting from coins.

Remove bad white spaces from JSON api documentation.