hdkeychain: Provide SerializedPubKey method
The SerializedPubKey method returns the internal and memoized bytes
representing the compressed serialization of the HD key's secp256k1
public key.  The motivation for providing this method is to avoid an
extra serialization and deserialization through *secp256k1.PublicKey
when creating the pubkey hash.  Wallet profiling revealed that roughly
30% of CPU time deriving child addresses from account branch extended
keys was a result of the extra forced serialization steps.