dcrutil: Provide privkey access for WIFs
This adds a PrivKey method to dcrutil.WIF which returns the serialized
private key.

While here and before a stable release of dcrutil/v3 has been tagged,
the method SerializePubKey is renamed to simply PubKey.
SerializePubKey was not an accurate method name, as it described an
action performed rather than the data returned, and this
implementation was returning already-serialized bytes.

A mention in a doc comment to the unexported field ecType (which is
probably a find & replace artifact) is removed.  Because the WIF might
be for a key that is not ECDSA (e.g. ed25519 or schnorr keys), the
ecType field is renamed to scheme.