hdkeychain:  ECPrivKey -> SerializedPrivKey.
This replace the ECPrivKey method with a SerializedPrivKey method to be
consistent with the newly added SerializePubKey method and updates the
tests accordingly.

The primary reason for this change is that it removes the secp256k1 type
from the public API and is more efficient in practice because the
serialized bytes are already available internally and callers are
exceedingly likely to just serialize the returned privkey again, so the
current approach  almost always resulted in an additional needless round
trip through parsing the private key.