hdkeychain: Remove ECPubKey.
This removes the ECPubKey method in favor of the newly added
SerializedPubKey method, updates the tests and example to use the new
method instead, and finally updates the README accordingly as well.

The primary reason for this change is that it removes the secp256k1 type
from the public API and is more efficient in practice because the
serialized bytes are already available internally and callers are
exceedingly likely to just serialize the returned pubkey again, so the
current approach  almost always resulted in an additional needless round
trip through parsing the public key, which is rather expensive.