hdkeychain: Use direct hashes and remove dcrutil dep
This change replaces usage of chainhash (which is intended solely for
transaction and blocks) and dcrutil.Hash160 (used in scripts for
hashing redeem scripts and pubkeys) with direct calls to each hashing
function from dcrd's crypto subpackages.  This improves clarity and
shortens the transitive package imports.  Notably, dcrutil and
everything it depends on is no longer transitively imported when
importing the hdkeychain package.