secp256k1: Use mod n scalar in signature type.
This updates the internal implementation of the Signature type to use
the new ModNScalar type instead of big.Ints and modifies NewSignature to
accept the new ModNScalar type accordingly.

Signature parsing and signing operations are now completely free of
big.Ints.  Verification still relies on them due to their use in public
keys.