secp256k1: Rework DER sig parsing tests.
This reworks the DER signature parsing tests to ensure 100% test
coverage of all branches in the parsing code and also test for the
explicit reason for a failure to ensure that each test is actually
testing the intended condition.

Compare this to the previous more generic validity testing which only
tested that a signature failed to parse as opposed to _why_ it failed to
parse.  The result was that some of the tests were not actually testing
what they claimed.