schnorr: Remove unused pubkey recovery bits.
This removes the code related to public key recovery from a schnorr
signature since it is not used, there are not currently any plans to use
it, and the existing code would require some effort to get it production
ready.

This code will still be in the repository history if it is ever needed
at some point in the future.