mempool: Correct MaybeAcceptDependents mutex.
The MaybeAcceptDependents function was added as a part of the logic to
track tickets with non-approved inputs separately in a staging pool.
Given it modifies the contents of the mempool, it needs to hold a write
mutex as opposed to a read mutex as otherwise it can modify the mempool
out from under readers.

While here, comment the required locking semantics on the other
functions that were added as a part of the staging logic as well.