secp256k1: Add AsJacobian method to pubkey.
This add a new method to the PublicKey type named AsJacobian which
converts the public key into a Jacobian point with Z=1.  This allows the
public key to be treated a Jacobian point in the secp256k1 group in
calculations.

It also adds an associated test to ensure proper functionality.