schnorr: Add error support for errors.Is/As.
This updates the Error and ErrorCode types to provide support for
errors.Is and errors.As and adds a full set of tests for the error
infrastructure so the error handling is more consistent with the rest of
the secp256k1 module.

While here, it removes the unused ErrPubKeyOffCurve error and renames
the error creation func to signatureError to be more consistent.