wire: Improve error handling.
This updates the error handling in the wire package to be more
consistent with the rest of the code base, tightens some of the error
handling in message parsing to harden the protocol, and adds tests to
ensure the error implementations works correctly with the standard
library errors.Is and errors.As functions introduced in Go 1.13.

In particular, the error code semantics have been changed so the first
error starts at 0 instead of treating 0 is a sentinel "no error" value
because it goes against Go best practices which designed the errors
specifically to avoid the use of sentinel values.  Further, the use of 0
as a sentinel means that there are two different ways to indicate "no
error", namely nil and an error code of zero.  Thus callers have to know
that they need to check for both conditions which is incredibly error
prone.  In fact, there is already at least one instance in the code base
that has that mistake in the tests.

Finally, while here, this hardens the protocol by enforcing more strict
parsing requirements lower in the software stack instead of relying on
upper layers to take care of it.