peer: Set a default idle timeout if not specified.
This modifies the peer initialization code to set a default idle timeout
of two minutes if the caller did not specify one.

This is desirable since the default is the zero value otherwise, which
results in an immediate disconnection for callers that don't specify a
timeout.

It is worth noting that the server code already has a default setting
that can be configured via the command-line or a configuration file
which is used to configure the peers, so this has no effect on dcrd, but
it is a nice addition for other consumers of the peer package.