connmgr: Tighten TestNetworkFailure.
This reworks the TestNetworkFailure test to more accurately test the
intended condition by signaling once the max number of failed attempts
has been reached prior to shutting down the connection manager and
checking the expected number of total dials attempted.

It also tightens the number of expected dials to a more accurate value
to better ensure an accurate result.