connmgr: Tighten TestStopFailed.
This reworks the TestStopFailed test to more accurately test the
intended condition without sleep hacks.  It does this by setting the
retry timeout to a known value, signaling exactly when a dial attempt
has been made, and responding to that signal by shutting down the
connection manager after precisely half the duration of the retry
timeout has elapsed.

Finally, it renames the test to TestShutdownFailedConns to more
accurately reflect what it is testing.