connmgr: Convert lifecycle to context.
This modifies the lifecycle of the connection manager to use the
expected pattern for running subsystems based on contexts.

In particular, this replaces the Start, Stop, and Wait methods with a
single method named Run and arranges for it to block until the provided
context is cancelled.  This is more flexible for the caller since it can
easily turn blocking code into async code while the reverse is not true.

The new Run method waits for all goroutines that it starts to shutdown
before returning to help ensure an orderly shutdown.

Since there are exported methods that send messages to the connection
handler goroutine via a channel and that goroutine is stopped during the
shutdown process, all sends select across both the channel in question
as well as a quit channel which is closed when the context it canceled.
This ensures callers can't end up blocking on send to a goroutine that
is no longer running without needing additional mutexes.

As part of the conversion, this adds various comments to better describe
the expected behavior and modifies timeout functions to select across
the quit channel as well to support cancellation.

Finally, all tests and callers are updated to use the new API and the
server is modified to take advantage of the new lifecycle semantics by
taking ownership of the connection manager and seeding code directly in
its Run method where it more logically makes sense.