rpcclient: Resurrect validateaddress and verifymessage support
These methods were accidentally removed from the client type when the
dcrwallet method support was removed from the package.  However, dcrd
also implements verifymessage and a less-featureful version of the
validateaddress method, so the client ought to provide direct support
for these.

Compared to the earlier version of this the validateaddress methods,
these reintroduced methods use dcrd's defined type rather than
depending on the dcrwallet definition.