rpcserver: Correct adaptor for utxo entry fetch.
This corrects the adaptor that converts a concrete blockchain.UtxoEntry
and rpcserver.UtxoEntry interface that was recently introduced in the
commit that created both it and the rpcserver.Chain interface.

In particular, it should not return an instance of the wrapper when
either the entry does not exist (is nil) or there was an error.