txscript: Add tests for new strict null data func.
This adds tests for the new IsStrictNullData function and also updates
the function to take a uint32 since it is going to be used consensus via
the upcoming decentralized treasury code and therefore types that vary
by architecture can't be used.

Finally, it also renames the expectedLength param to requiredLen to make
it more clear it is a requirement.