rpcserver: Remove global config dependency.
This removes the dependency that rpcserver has on the global config in
the main module.  It accomplishes this by passing the necessary config
information through the rpcserverConfig struct.  Additionally, this adds
a Lookup method to the rpcserver.ConnManager interface.

This is part of the effort to move rpcserver into a separate internal
package.