rpcserver: Remove PeerNotifier dependency.
This removes the dependency that rpcserver has on the PeerNotifier
interface defined in blockmanager.go.  It accomplishes this by removing
the call to AnnounceNewTransactions and instead calling the already
defined ConnMgr.RelayTransactions interface method, followed by directly
calling NotifyNewTransactions to notify websocket clients of the newly
accepted transactions.

This is part of the effort to move rpcserver into a separate internal
package.