rpcserver: Export rpcwebsocket Notify functions.
This adds the following exported RPCServer functions, which delegate to
the corresponding underlying rpcwebsocket notification functions:
  - NotifyWork
  - NotifyStakeDifficulty
  - NotifyNewTickets
  - NotifyBlockConnected
  - NotifySpentAndMissedTickets
  - NotifyBlockDisconnected
  - NotifyReorganization
  - NotifyWinningTickets

This is done so that blockmanager.go and server.go can still access
these functions after rpcserver is moved to a separate internal package.

This is part of the effort to move rpcserver into a separate internal
package.