mempool: Store staged transactions as TxDesc
This change modifies the mempool to store staged transactions
as a *TxDesc rather than a *dcrutil.Tx. The advantage of doing
this is to avoid recalculating transaction types through
stake.DetermineTxType since the type would have been captured
in a TxDesc when the transaction initially entered the mempool.

Additionally, rather than returning an array of redeemers for a
given transaction, a function provided by the caller is invoked
for each redeemer. This eliminates unnecessary allocations by
allowing the caller to access each redeemer without needing to
rely on an intermediary buffer.