main: Simplify startup logic slightly.
This modifies the main startup logic to make use of the fact that the
server now has its lifecycle tied to the same context that is cancelled
when a shutdown signal is received.

The end result is that the separate goroutine it was running in before
and associated synchronization logic is no longer necessary.